Add message view page and improve message handling
Introduces a new NumberHelper for formatting large numbers and adds a dedicated view page for messages with copy/share functionality. Updates validation in MessageController, changes the 'limit' column to unsignedBigInteger, improves error handling and form persistence, and updates routes to support the new view. Also enhances UI with cache-control headers and error display.

+++ database/migrations/2025_09_18_071913_create_messages_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->bigIncrements('id');
             $table->string('token', 64)->unique();
             $table->string('message', 150);
-            $table->unsignedInteger('limit')->default(1);
+            $table->unsignedBigInteger('limit')->default(1);
             $table->unsignedBigInteger('expires_at');
             $table->timestamps();
             $table->softDeletes();
